Workaround for actual Weatherflow API bug
Currently in the returned JSON data from the Weatherflow API the `current_conditions->icon` has a `cc-` prefix. This caused that the weather type in Loxone was `0` and the new app displayed only a black screen. The black screen problem was solved by a Loxone engineer and the root cause is solved by removing the `cc-`. I also opened a ticket at the WeatherFlow Support to understand, if this is a API bug or if the documentation (https://apidocs.tempestwx.com/reference/get_better-forecast-1) was not updated.